wit example cleanup
