+++ azure_devops_rust_api/Cargo.toml
@@ -233,11 +233,11 @@ name = "wiki_pages_create_or_update"
 required-features = ["wiki"]
 
 [[example]]
-name = "wit_create_work_item"
+name = "wit_work_item_create"
 required-features = ["wit"]
 
 [[example]]
-name = "wit"
+name = "wit_work_item_get"
 required-features = ["wit"]
 
 [[example]]
