+++ azure_devops_rust_api/run_all_examples
@@ -108,6 +108,6 @@ cargo run --example search_work_item --features="search" $WORK_ITEM_SEARCH_TEXT
 cargo run --example test_runs_list --features="test"
 cargo run --example client_pipeline_policy --features="git"
 cargo run --example wiki_pages_create_or_update --features="wiki" $ADO_WIKI_ID $ADO_WIKI_PAGE_PATH $ADO_WIKI_CONTENT
-cargo run --example wit_create_work_item --features="wit"
+cargo run --example wit_work_item_create --features="wit"
 
 echo "Done"
